+++ NEWS.rst
@@ -145,9 +145,10 @@ Changes
 * All built/installed Python modules are now in the ``zstandard``
   package. Previously, there were modules in other packages. (#115)
 * C source code is now automatically formatted with ``clang-format``.
-* ``ZstdCompressor.stream_writer()`` now accepts a ``closefd`` argument
-  to control whether the underlying stream should be closed when the
-  ``ZstdCompressionWriter`` is closed.
+* ``ZstdCompressor.stream_writer()`` and ``ZstdDecompressor.stream_writer()``
+  now accept a ``closefd`` argument to control whether the underlying
+  stream should be closed when the ``ZstdCompressionWriter`` or
+  ``ZstdDecompressionWriter`` are closed.
 
 0.14.1 (released 2020-12-05)
 ============================